third_party/blink/web_tests/flag-specific/disable-layout-ng/external/wpt/css/cssom-view/cssom-getClientRects-002-expected.txt

This is a testharness.js-based test.
FAIL CSSOM View - GetClientRects().length is the same regardless source new lines assert_equals: count2 expected 1 but got 2
Harness: the test ran to completion.
